chore: refactor push certificate tests and cleanup regtest
