+++ crates/topos/src/components/regtest/mod.rs
@@ -6,13 +6,12 @@ use tokio::{
     sync::{mpsc, oneshot},
 };
 use topos_certificate_spammer::CertificateSpammerConfig;
-use tracing::{debug, error, info};
+use tracing::{error, info};
 use tracing_opentelemetry::OpenTelemetrySpanExt;
 
 use crate::tracing::setup_tracing;
 
 pub(crate) mod commands;
-pub(crate) mod services;
 
 pub(crate) async fn handle_command(
     RegtestCommand {
@@ -21,28 +20,6 @@ pub(crate) async fn handle_command(
     }: RegtestCommand,
 ) -> Result<(), Box<dyn std::error::Error>> {
     match subcommands {
-        Some(RegtestCommands::PushCertificate(cmd)) => {
-            _ = setup_tracing(verbose, false, None, None)?;
-            debug!("Start executing PushCertificate command");
-            match services::push_certificate::check_delivery(
-                cmd.timeout_broadcast,
-                cmd.format,
-                cmd.nodes,
-                cmd.timeout,
-            )
-            .await
-            .map_err(Box::<dyn std::error::Error>::from)
-            {
-                Err(e) => {
-                    error!("Check failed: {:?}", e);
-                    std::process::exit(1);
-                }
-                _ => {
-                    info!("Check passed");
-                    Ok(())
-                }
-            }
-        }
         Some(RegtestCommands::Spam(cmd)) => {
             let config = CertificateSpammerConfig {
                 target_nodes: cmd.target_nodes,
