--- crates/topos/src/components/regtest/services.rs
@@ -1,1 +1,0 @@
-pub(crate) mod push_certificate;
